feat: phase-17 core production qualification
36-entry matrix (3 themes x 3 presets x seeds 0/42/99/255): generator leak
fixes (pit-omission lights, entity void origins, spawn clearance/hull-2
centering), completed vertical cost reservation, 36/36 compile warning-free/
leak-free/strict-reload + spatial witnesses; frozen manifest with real
hashes, second-pass byte-identical; overflow hardening (props/lighting
modular seed indices); qualification suite 55 tests + request fixture
inventory + compatibility tests; cross-arch identity digest CI gate;
6 deterministic 1280x720 captures + vision review (theme/portal/route/
lighting/material PASS); grand-volume arena captures via new
--acceptance-camera-origin/--look-at CLI; --wsi-lifecycle-test
(resize/minimize/restore PASS live); runtime traversal 20/20; live WSI
startups PASS; release audit clean; fixed real bugs: package.rs
DirectImportStaging 16->8 copy panic, vertical suite fixtures now carry
brush-modeled drop volumes; qualification.json 22 PASS / 2 disclosed
blocked rows (ssim-reference-renderer external assets, prop-visibility
inconclusive).

+++ src/bsp_runtime/src/package.rs
@@ -1199,7 +1199,7 @@ impl DirectImportStaging {
             })?
             .as_nanos();
         let pid = std::process::id();
-        rng[..8].copy_from_slice(&nanos.to_le_bytes());
+        rng[..8].copy_from_slice(&(nanos as u64).to_le_bytes());
         rng[8..12].copy_from_slice(&pid.to_le_bytes());
         rng[12..16].copy_from_slice(b"\xDE\xAD\xBE\xEF");
         let hex = rng.iter().map(|b| format!("{:02x}", b)).collect::<String>();
